--- src/custom.scss
@@ -1,5 +1,0 @@
-$theme-colors: (
-    "primary": #1D7874
-);
-
-@import "../node_modules/bootstrap/scss/bootstrap";
