+++ src/styles/custom.scss
@@ -1,0 +1,5 @@
+$theme-colors: (
+    "primary": #1D7874
+);
+
+@import "../../node_modules/bootstrap/scss/bootstrap";
